+++ react/src/lib.rs
@@ -1,146 +1,13 @@
-/// `InputCellID` is a unique identifier for an input cell.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct InputCellID(usize);
-/// `ComputeCellID` is a unique identifier for a compute cell.
-/// Values of type `InputCellID` and `ComputeCellID` should not be mutually assignable,
-/// demonstrated by the following tests:
-///
-/// ```compile_fail
-/// let mut r = react::Reactor::new();
-/// let input: react::ComputeCellID = r.create_input(111);
-/// ```
-///
-/// ```compile_fail
-/// let mut r = react::Reactor::new();
-/// let input = r.create_input(111);
-/// let compute: react::InputCellID = r.create_compute(&[react::CellID::Input(input)], |_| 222).unwrap();
-/// ```
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct ComputeCellID(usize);
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct CallbackID(usize);
+pub mod ids;
+pub use ids::{CallbackID, CellID, ComputeCellID, InputCellID};
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub enum CellID {
-    Input(InputCellID),
-    Compute(ComputeCellID),
-}
-
-impl CellID {
-    fn idx(&self) -> usize {
-        match self {
-            Self::Input(InputCellID(idx)) => *idx,
-            Self::Compute(ComputeCellID(idx)) => *idx,
-        }
-    }
-}
-
-impl From<InputCellID> for CellID {
-    fn from(id: InputCellID) -> CellID {
-        Self::Input(id)
-    }
-}
-
-impl From<ComputeCellID> for CellID {
-    fn from(id: ComputeCellID) -> CellID {
-        Self::Compute(id)
-    }
-}
+pub mod cells;
+use cells::{Cell, ComputeCell, InputCell};
 
 #[derive(Debug, PartialEq)]
 pub enum RemoveCallbackError {
     NonexistentCell,
     NonexistentCallback,
-}
-
-struct InputCell<T> {
-    value: T,
-    fwd: Vec<ComputeCellID>,
-}
-
-impl<T> InputCell<T> {
-    fn new(value: T) -> Self {
-        Self {
-            value,
-            fwd: Vec::new(),
-        }
-    }
-}
-
-struct ComputeCell<T> {
-    dependencies: Vec<CellID>,
-    computation: Box<dyn Fn(&[T]) -> T>,
-    cache: T,
-    fwd: Vec<ComputeCellID>,
-}
-
-impl<T> ComputeCell<T>
-where
-    T: Copy + PartialEq,
-{
-    /// caution: use only when you know all dependencies are legal
-    fn calculate<F>(cells: &[Cell<T>], dependencies: &[CellID], computation: F) -> T
-    where
-        F: Fn(&[T]) -> T,
-    {
-        let values: Vec<T> = dependencies
-            .iter()
-            .map(|cid| cells[cid.idx()].value())
-            .collect();
-        computation(&values)
-    }
-
-    fn new<F>(reactor: &Reactor<T>, dependencies: &[CellID], computation: F) -> Result<Self, CellID>
-    where
-        F: 'static + Fn(&[T]) -> T,
-    {
-        // ensure that all dependencies are legal
-        if let Some(missing) = dependencies
-            .iter()
-            .find(|id| id.idx() >= reactor.cells.len())
-        {
-            return Err(*missing);
-        }
-
-        Ok(Self {
-            dependencies: dependencies.to_owned(),
-            cache: Self::calculate(&reactor.cells, dependencies, &computation),
-            computation: Box::new(computation),
-            fwd: Vec::new(),
-        })
-    }
-
-    fn recompute(&mut self, cells: &[Cell<T>]) {
-        self.cache = Self::calculate(cells, &self.dependencies, &self.computation);
-    }
-}
-
-enum Cell<T> {
-    Input(InputCell<T>),
-    Compute(ComputeCell<T>),
-}
-
-impl<T: Copy + PartialEq> Cell<T> {
-    fn fwd(&self) -> &[ComputeCellID] {
-        match self {
-            Self::Input(input) => &input.fwd,
-            Self::Compute(compute) => &compute.fwd,
-        }
-    }
-
-    fn fwd_mut(&mut self) -> &mut Vec<ComputeCellID> {
-        match self {
-            Self::Input(input) => &mut input.fwd,
-            Self::Compute(compute) => &mut compute.fwd,
-        }
-    }
-
-    fn value(&self) -> T {
-        match self {
-            Self::Input(ic) => ic.value,
-            Self::Compute(cc) => cc.cache,
-        }
-    }
 }
 
 #[derive(Default)]
